Push Details

+++ src/operations/delete_all_nominee.php
@@ -15,15 +15,25 @@
     $dt = date('Y-m-d G:i:s');
     $action = "Removed all candidates from "."$campus"." campus";
 
-    $query2 = "INSERT INTO tblogs(name,action,timestamp) VALUES('$tempCampus', '$action','$dt')";
-    $query3 = "DELETE from tb_votes where campus = '$campus'";
+    $query4 = "Select admin_id from tbadmin where username = '$tempCampus'";
+    $result = mysqli_query($conn, $query4);
+    $row = mysqli_fetch_assoc($result);
+    if ($row) {
+    $admin_id = $row['admin_id'];
+        $query2 = "INSERT INTO tb_admin_action_logs(admin_id,action,log_action_date) VALUES('$admin_id', '$action','$dt')";
+        $query3 = "DELETE from tb_votes where campus = '$campus'";
 
-    if(mysqli_query($conn,$query) == TRUE){
-        $success = 'true';
-        mysqli_query($conn,$query2);
-        mysqli_query($conn,$query3);
-    }else{
-        $statusMsg = "Unable to delete all nominees";
+        if(mysqli_query($conn, $query)){
+            $success = 'true';
+            mysqli_query($conn,$query2);
+            mysqli_query($conn,$query3);
+        }else{
+            $statusMsg = "Unable to delete all nominees";
+        }
+    } else {
+    echo "Admin not found.";
     }
+
+
 
 ?>
