+++ src/operations/delete_position.php
@@ -14,15 +14,23 @@
     $dt = date('Y-m-d G:i:s');
     $action = "Removed local council position "."\"".$position."\"";
 
-    $query2 = "INSERT INTO tblogs(name,action,timestamp) VALUES('$tempCampus', '$action','$dt')";
+$query4 = "Select admin_id from tbadmin where username = '$tempCampus'";
+$result = mysqli_query($conn, $query4);
+$row = mysqli_fetch_assoc($result);
+if ($row) {
+    $admin_id = $row['admin_id'];
+    $query2 = "INSERT INTO tb_admin_action_logs(admin_id,action,log_action_date) VALUES('$admin_id', '$action','$dt')";
 
     $query = "DELETE from tbposition where code = '$id'";
 
-    if(mysqli_query($conn,$query) == TRUE){
+    if (mysqli_query($conn, $query) == TRUE) {
         $success = 'true';
-        mysqli_query($conn,$query2);
-    }else{
+        mysqli_query($conn, $query2);
+    } else {
         $statusMsg = "Unable to delete position";
     }
+}else {
+
+}
 
 ?>
